Introducing the gemification of filterable_lib
Why ?
Imho, there are several very good reasons that justify the usage of a gem, even if it's an intern gem, instead of a cross-repo lib.
The first one, and most important, is the maintanability of the functionnality. A lib repo doesn't have it's own stack and dependencies. Trying to import it into a new repo could (and will) lead
to many compatibility issues, different version, different tools used.
Also, writing specs to a lib folder is kinda weird and would require lots of custom things, while gem are meant to be tested and isolated.

Last argument toward this is the easiness of using a gem, vs a lib folder to copy paste.
A gem is meant to be used across repository, and easy to replicate, with generators for instance, and doesn't rely on copy-pasta.

How ?
I took lots of inspiration on a similar lib ([toschas/filterable](https://github.com/toschas/filterable)) to get ideas of code organization. This commit is far from being the final one toward
having a really working gem. More re-organisation of items, components, and a lot more of specs :nerd-face: